Cambios en el log y consulta
